Fix #158: add missing hover-color case
#33 introduced an update to hover-color() that I think is missing the
force-use-color parameter.

When I have a style such as:

li {
background-color transparent // or initial, to reset a
background-color from above
hover-color red true
}

The result is a list item whose background swaps to red on hover,
rather than respecting the second argument of true to hover-color().

Adding a force-use-color check fixes this.